refactor: fetcher 분리
