+++ src/api/search/entity.ts
@@ -1,3 +1,13 @@
 export interface SearchQueryParams {
   searchQuery : string;
 }
+
+export interface ShopsParams {
+  searchQuery: string;
+  location: Coords;
+}
+
+export interface Coords {
+  latitude: number,
+  longitude: number
+}
