+++ src/pages/SearchDetails/hooks/useFetchShops.ts
@@ -1,0 +1,24 @@
+import useGeolocation from 'utils/hooks/useGeolocation';
+import { useQuery } from 'react-query';
+import { fetchShops } from 'api/search';
+import { ShopsParams } from 'api/search/entity';
+
+const useFetchShops = (searchQuery: string) => {
+  const options = {
+    maximumAge: 1000,
+  };
+  const { location } = useGeolocation(options);
+  const params = { searchQuery, location };
+  const {
+    isLoading, isError, data, refetch,
+  } = useQuery('shop', () => fetchShops(params as ShopsParams), { enabled: !!location });
+
+  const isFetching = isLoading || !(location);
+  const shops = data?.data.content;
+
+  return {
+    isFetching, isError, data: shops, refetch,
+  };
+};
+
+export default useFetchShops;
